requirements.txt
Flask==2.3.2
Werkzeug==2.3.4
twilio==8.2.1
googletrans==3.0.0
PyPDF2==3.0.1
pinecone-client==2.2.1
langchain==0.0.174
openai
python-dotenv
gunicorn
boto3
scikit-learn
tiktoken
unstructured
pdf2image
azure.core